integrations/python/dataloader/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openhouse.dataloader"
version = "0.0.1"
description = "A Python library for distributed data loading of OpenHouse tables"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["datafusion==51.0.0", "pyiceberg~=0.11.0", "opentelemetry-api>=1.20.0"]

[project.optional-dependencies]
dev = ["ruff>=0.9.0", "pytest>=8.0.0", "opentelemetry-sdk>=1.20.0"]

[tool.hatch.build.targets.wheel]
packages = ["src/openhouse"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.format]
quote-style = "double"
